Update LICENSE and refactor utils
