+++ src/commands/add/mod.rs
@@ -1,6 +1,6 @@
-use std::path::PathBuf;
 use crate::errors;
 use crate::utils;
+use std::path::PathBuf;
 
 pub fn handle_command(path: PathBuf, hosts: Vec<String>) -> errors::RhostmanResult<()> {
     let content = utils::hosts_to_string(path)?;
